Fixed black hole projectile affecting wrong maps
Right now, it's using Find.CurrentMap, which is the map the player is on - not the projectile. This allowed this projectile to affect different maps instead of the one they were spawned into by just switching the map the player is on.

This fixes this issue by using the actual map it's located in.